requirements.txt
Django==3.0.8
djangorestframework==3.11.0
gunicorn==20.0.4
python-dateutil==2.8.1
